demo05/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
